Seperate modecule for mapping over radio btn
