+++ components/atom/button.jsx
@@ -41,3 +41,14 @@ export const submitButton = (handleSubmit, Name) => {
         </div>
     );
 };
+
+export const RadioButton = ({ value, checked, onChange }) => (
+    <>
+        <input
+            type="radio"
+            onChange={onChange}
+            name="radio"
+            defaultChecked={checked}
+        />
+    </>
+);
