--- components/atom/label.jsx
@@ -1,1 +1,0 @@
-export const label = (name) => <label>{name}</label>;
