--- components/atom/radio.jsx
@@ -1,10 +1,0 @@
-export const RadioButton = ({ value, checked, onChange }) => (
-    <>
-        <input
-            type="radio"
-            onChange={onChange}
-            name="radio"
-            defaultChecked={checked}
-        />
-    </>
-);
